requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
cm-api==19.2.2
distlib==0.3.4
Django==4.0.2
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.4.2
frozenlist==1.3.0
future==0.18.2
hdfs==2.6.0
idna==3.3
install==1.3.5
isodate==0.6.1
lxml==4.7.1
multidict==6.0.2
mysql-connector-python==8.0.28
numexpr==2.8.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
platformdirs==2.4.1
protobuf==3.19.4
PyHive==0.6.4
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tables==3.7.0
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.1
yarl==1.7.2
zeep==4.1.0
